Hook(useFetchProducts):add customHoooks useFetchProducts
